Remove `Federated` Datatype and `FederatedObject`
