+++ src/main/java/org/tugraz/sysds/common/Types.java
@@ -38,16 +38,13 @@ public class Types
 	 * Data types (tensor, matrix, scalar, frame, object, unknown).
 	 */
 	public enum DataType {
-		FEDERATED, TENSOR, MATRIX, SCALAR, FRAME, LIST, UNKNOWN;
+		TENSOR, MATRIX, SCALAR, FRAME, LIST, UNKNOWN;
 		
 		public boolean isMatrix() {
 			return this == MATRIX;
 		}
 		public boolean isTensor() {
 			return this == TENSOR;
-		}
-		public boolean isFederated() {
-			return this == FEDERATED;
 		}
 		public boolean isFrame() {
 			return this == FRAME;
